Exporting Dataset Functions

OPFLearn.save_polytope — Method

Saves the A and b matrices defining the polytope Ax < b to two seperate comma delimited csv files.

OPFLearn.save_results_csv — Method

Saves the AC_output, AC_input, duals to a csv file

OPFLearn.save_stats — Method

Saves the last iteration of level 1 stats to a csv file. If the csv file does not already exist then calls save_stats to save the stats data with a header to a new csv file.

Managing Results

OPFLearn.results_to_array — Method

Takes the results objects from a create samples run and converts them to a single array. Requires the save_order, an array of variables, to save from the results objects. Takes optional boolean arguments header, whether to include a header as the first row, and imag_j, whether to convert complex data values to strings and replace 'im' with 'j' for datasets that are to be used outside of the Julia environment.